refactor(mcp): move server instructions to instructions.ts so codegen can import them without the SDK
